% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insider.R
\name{fit.insider}
\alias{fit.insider}
\title{fit insider object}
\usage{
fit.insider(object, latent_dimension = NULL, lambda = NULL, alpha = NULL)
}
\arguments{
\item{object}{an insider object}

\item{latent_dimension}{the rank of latent representations}

\item{lambda}{l2-regularization for insider}

\item{alpha}{l1-regularization for insider}
}
\value{
a fitted insider object
}
\description{
fit insider object
}
