% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insider.R
\name{fit}
\alias{fit}
\title{fit insider object with selected hyperparameter.}
\usage{
fit(
  object,
  latent_dimension = NULL,
  lambda = NULL,
  alpha = NULL,
  partition = 0
)
}
\arguments{
\item{object}{an insider object}

\item{latent_dimension}{the selected rank of latent representations}

\item{lambda}{a vector of l2 penalty}

\item{alpha}{a vector of l1 penalty}

\item{partition}{should be either 1 or 0. 1 indicates only trainset used for fitting, and 0 indicates the whole data matrix used for fitting.}
}
\value{
the fitted insider object with the provided hyperparameters.
}
\description{
fit insider object with selected hyperparameter.
}
\examples{
fit(object, latent_dimension = as.integer(num_factors), lambda = lambda, alpha = alpha, partition = 0)
}
