\name{insider-package}
\alias{insider-package}
\alias{insider}
\docType{package}
\title{
\packageTitle{insider}
}
\description{
\packageDescription{insider}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{insider}
\packageIndices{insider}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{insider}

Maintainer: \packageMaintainer{insider}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
