% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insider.R
\name{insider}
\alias{insider}
\title{create an insider object with the provided parameters}
\usage{
insider(
  data,
  confounder,
  interaction_idx,
  split_ratio = 0.1,
  global_tol = 1e-09,
  sub_tol = 1e-05,
  tuning_iter = 30,
  max_iter = 50000
)
}
\arguments{
\item{data}{the data for insider interpretation}

\item{confounder}{a matrix of dummy variables, which indicates the belonging of each row for several covariates. Here covariates are discreate variables, such as gender, sex, development stages, and etc.}

\item{interaction_idx}{a integer vector of indices corresponding to confounder matrix to incorporate interaction between covariates, e.g., as.integer(c(1, 2)) to introduce interactions between the first and second covariates}

\item{split_ratio}{a proportion of elements from data that will be considered as testset, and the left over is testset.}

\item{global_tol}{the global convergence criteria. When the criteria is met, iteraction will be terminated.}

\item{sub_tol}{the convergence creteria for elastic net problems, which can be decreased with decreasing the loss of the objective function.}

\item{tuning_iter}{number of steps will run in tuning}

\item{max_iter}{the maxiumme number of iterations. When it is reached, iteraction will terminated even if the global convergence criteria does not meet.}
}
\value{
an insider object
}
\description{
create an insider object with the provided parameters
}
\examples{
object <- insider(data, as.matrix(confounders), as.integer(c(1,2)), global_tol = 1e-10)

}
