% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ratio_splitter}
\alias{ratio_splitter}
\title{matrix splitting via element-wise sampling without replacement}
\usage{
ratio_splitter(data, ratio = 0.1, rm.na.col = T)
}
\arguments{
\item{data}{data matrix for splitting}

\item{ratio}{a proportion of elements considered as testset}

\item{rm.na.col}{whether remove columns with all zeros}
}
\value{
a list
}
\description{
matrix splitting via element-wise sampling without replacement
}
\examples{
results <- ratio_splitter(data)
}
