% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insider.R
\name{tune.insider}
\alias{tune.insider}
\title{tune hyperparameters for insider object}
\usage{
tune.insider(object, latent_dimension = NULL, lambda = 1, alpha = 0.1)
}
\arguments{
\item{object}{an insider object}

\item{latent_dimension}{the rank of latent representations}

\item{lambda}{l2-regularization for insider}

\item{alpha}{l1-regularization for insider}
}
\value{
tuning results
}
\description{
tune hyperparameters for insider object
}
