% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insider.R
\name{insider}
\alias{insider}
\title{create an insider object with the provided parameters}
\usage{
insider(
  data,
  confounder,
  split_ratio = 0.1,
  global_tol = 1e-09,
  sub_tol = 1e-05,
  tuning_iter = 30,
  max_iter = 50000
)
}
\arguments{
\item{data}{the whole data for insider, including trainset and testset}

\item{confounder}{a matrix with each column indicating the belonging for each covariate}

\item{split_ratio}{a proportion of elements from data considered as testset}

\item{global_tol}{the global convergence creteria}

\item{sub_tol}{the convergence creteria for each elastic net problem}

\item{tuning_iter}{number of steps will run in tuning}

\item{max_iter}{maxiumme number of steps will run in fitting insider}
}
\value{
an insider object
}
\description{
create an insider object with the provided parameters
}
\examples{
..
}
